ENH: Added support for optional terms.
The entries in the model "UserTermsAndConditions" have now a
different meaning. If a row exists, this means "the terms have been
presented to the user and he/she decided to accept or not to accept"
or in short "have been seen".

If the field "date_accepted" is not NULL, the terms have been
accepted. If it is NULL, the user decided not to accept it.

The latter is only possible for "optional terms", these are instances of the
model "TermsAndConditions" having "boolean=True" (new field).